remove(async)!: Client
It's just an ApiClient meow
